// condoplanner-app/src/apiClient/models/BlockDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ApartmentDto } from './ApartmentDto';
import {
    ApartmentDtoFromJSON,
    ApartmentDtoFromJSONTyped,
    ApartmentDtoToJSON,
    ApartmentDtoToJSONTyped,
} from './ApartmentDto';

/**
 * 
 * @export
 * @interface BlockDto
 */
export interface BlockDto {
    /**
     * 
     * @type {number}
     * @memberof BlockDto
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof BlockDto
     */
    name?: string | null;
    /**
     * 
     * @type {Array<ApartmentDto>}
     * @memberof BlockDto
     */
    apartments?: Array<ApartmentDto> | null;
}

/**
 * Check if a given object implements the BlockDto interface.
 */
export function instanceOfBlockDto(value: object): value is BlockDto {
    return true;
}

export function BlockDtoFromJSON(json: any): BlockDto {
    return BlockDtoFromJSONTyped(json, false);
}

export function BlockDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): BlockDto {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'apartments': json['apartments'] == null ? undefined : ((json['apartments'] as Array<any>).map(ApartmentDtoFromJSON)),
    };
}

export function BlockDtoToJSON(json: any): BlockDto {
    return BlockDtoToJSONTyped(json, false);
}

export function BlockDtoToJSONTyped(value?: BlockDto | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'apartments': value['apartments'] == null ? undefined : ((value['apartments'] as Array<any>).map(ApartmentDtoToJSON)),
    };
}