// condoplanner-app/src/apiClient/models/EStatusVinculoResidencial.ts
/* tslint:disable */
/* eslint-disable */
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const EStatusVinculoResidencial = {
    Pendente: 'Pendente',
    Aprovado: 'Aprovado',
    Rejeitado: 'Rejeitado',
    Cancelado: 'Cancelado'
} as const;
export type EStatusVinculoResidencial = typeof EStatusVinculoResidencial[keyof typeof EStatusVinculoResidencial];


export function instanceOfEStatusVinculoResidencial(value: any): boolean {
    for (const key in EStatusVinculoResidencial) {
        if (Object.prototype.hasOwnProperty.call(EStatusVinculoResidencial, key)) {
            if (EStatusVinculoResidencial[key as keyof typeof EStatusVinculoResidencial] === value) {
                return true;
            }
        }
    }
    return false;
}

export function EStatusVinculoResidencialFromJSON(json: any): EStatusVinculoResidencial {
    return EStatusVinculoResidencialFromJSONTyped(json, false);
}

export function EStatusVinculoResidencialFromJSONTyped(json: any, ignoreDiscriminator: boolean): EStatusVinculoResidencial {
    return json as EStatusVinculoResidencial;
}

export function EStatusVinculoResidencialToJSON(value?: EStatusVinculoResidencial | null): any {
    return value as any;
}

export function EStatusVinculoResidencialToJSONTyped(value: any, ignoreDiscriminator: boolean): EStatusVinculoResidencial {
    return value as EStatusVinculoResidencial;
}